Test lookup in editor use case, move streaming tests into separate case.
